Bugfix in correction process

+++ Solution/Engine/AutoCorrector/AutoCorrector/CorrectionChecker.cs
@@ -63,11 +63,7 @@ public class CorrectionChecker
     }
     public async Task<string> MakeUnitTests(string req, string studentName,  string requirement, string function)
     {
-        if (studentName.Contains("Patrasc"))
-        {
-            return "Could not make unit test!";
-        }
-
+        
         LLMManager llmManager = new LLMManager();
         var result = await llmManager.WriteUnitTests(requirement, function);
         result = result.Replace("```", "");
@@ -82,7 +78,14 @@ public class CorrectionChecker
         await File.WriteAllTextAsync(unitTestFile, result);
 
         await CompileAsync(unitTestFile);
-        
+
+        var filePath = $"{Settings.UnitTestsPath}\\temp.exe";
+
+        if (!File.Exists(filePath))
+        {
+            return "Could not make unit test!";
+        }
+
         string output = await ExecuteAsync($"\"{Settings.UnitTestsPath}\\temp.exe\"");
 
         File.Delete($"{Settings.UnitTestsPath}\\temp.exe");
